requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain ecosystem
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.1
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1
langchain-core>=0.1.0

# Vector store
chromadb>=0.4.0

# Document processing
pymupdf>=1.23.0

# Embeddings
sentence-transformers>=2.2.0

# Type checking
typing-extensions>=4.0.0
